+++ src/main/java/by/vladzuev/locationreceiver/protocol/apel/decoder/location/locationdecoder/ApelShortLocationDecoder.java
@@ -12,7 +12,7 @@ public final class ApelShortLocationDecoder extends ApelLocationDecoder {
     }
 
     @Override
-    protected Double readHdop(final ByteBuf buffer) {
+    protected Byte readHdop(final ByteBuf buffer) {
         return null;
     }
